third_party/blink/web_tests/editing/pasteboard/paste-plaintext-user-select-none-expected.txt

This tests that you can paste into a plain-text editable area even if the body has user-select:none set on it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sample.textContent is "Test Passed"
PASS successfullyParsed is true

TEST COMPLETE


To test manually, copy "Test Passed" into clipboard, and paste in below box.

Test Passed
